semesterm4/practice/requirements.txt
python-telegram-bot
argostranslate
